move BOOT_POOL_NAME_VALID
